데이터 수정, 코드 작성
코인 추가 함수, Level Up 때 스테이트 보이게 하기, UI 추가

+++ Assets/Script/DataScript/TotalData.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class TotalData : ScriptableObject
 {
     [Header("# Total Coin")]
-    public int total_Coin;
+    public float total_Coin;
 
     [Header("# Player Upgrade")]
     public int[] hp_up;
